requirements.txt
Django==1.5.1
Markdown==2.3.1
Pillow==2.0.0
South==0.8.1
django-extensions==1.1.1
django-facebook==5.2.6
django-markitup==1.0.0
django-taggit==0.10a1
django-taggit-templatetags==0.4.6dev
django-templatetag-sugar==0.1
wsgiref==0.1.2
django-markitup
markdown
celery
